import additional lato font weights, move filter options up
