+++ package.json
@@ -34,6 +34,7 @@
     "react-router-redux": "^4.0.1",
     "react-select": "1.0.0-rc.5",
     "react-show-more": "^1.1.1",
+    "react-virtualized": "^9.9.0",
     "recharts": "^0.18.0",
     "redux": "^3.3.1",
     "redux-logger": "^2.6.1",
