+++ src/components/Catalog/@Catalog.scss
@@ -71,7 +71,7 @@
 }
 
 .Catalog__record-meta-value {
-  font-weight: lighter;
+  font-weight: 300;
   font-size: 13px;
   margin-left: 5px;
 }
@@ -91,7 +91,7 @@
 }
 
 .App--font-thin {
-  font-weight: lighter;
+  font-weight: 300;
 }
 
 .ais-RefinementList__showMoreDisabled {
@@ -107,13 +107,13 @@
 }
 
 .ais-RefinementList__itemLabel {
-  font-weight: lighter;
+  font-weight: 300;
   font-size: 13px;
 }
 
 .ais-RefinementList__itemCount {
   color: #222;
-  font-weight: lighter;
+  font-weight: 300;
 }
 
 .ais-Stats__root {
@@ -134,7 +134,7 @@
   border-radius: 0;
   background-color: transparent;
   color: #222;
-  font-weight: lighter;
+  font-weight: 300;
   padding: 4px 2px;
   font-size: 13px;
   margin-right: 5px;
