+++ src/containers/@containers.scss
@@ -72,7 +72,6 @@
   bottom: 0;
   top: 240px;
   overflow: scroll;
-  padding-right: 35px;
 }
 
 .VizContainer__stage {
